pyproject.toml
[project]
name = "chatwork-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.36.0",
    "requests>=2.32.5",
]
